Fix to refactor sources.
